Backend Complete

+++ backend/controllers/goalcontroller.js
@@ -1,10 +1,11 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../model/goalModel");
+const User = require("../model/UserModel");
 //@desc Get goals
 //@route GET /api/goals
 //@access Private
 const getgoals = asyncHandler(async (req, res) => {
-  const goal = await Goal.find();
+  const goal = await Goal.find({ user: req.user.id });
   res.status(200).json(goal);
 });
 
@@ -18,6 +19,7 @@ const setgoal = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
@@ -30,6 +32,17 @@ const updategoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal now found");
+  }
+  const user = await User.findById(req.user.id);
+  //Check for User
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  //Make sure the logged in user matched the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorised");
   }
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
@@ -45,6 +58,17 @@ const deletegoals = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal Does not exists");
+  }
+  const user = await User.findById(req.user.id);
+  //Check for User
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  //Make sure the logged in user matched the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorised");
   }
   await Goal.findByIdAndRemove(req.params.id);
   res.status(200).json({ id: req.params.id });
